修改机会点统计

+++ emucoo-dto/src/main/java/com/emucoo/dto/modules/report/ReportWorkVo.java
@@ -12,7 +12,7 @@ public class ReportWorkVo {
     private String executorHeadImgUrl;
     private Long executorID;
     private String executorName;
-    private Integer passRate;
+    private Float passRate;
     private String taskTitle;
     private String workID;
     private Integer workType;
